It seems to work a lot better now. Known existing issues: weights are a bit too low, activity not sufficient (?), analysis values are wrong (?)

+++ synaptogenesis/topographic_map_formation.py
@@ -327,8 +327,8 @@ else:
                                 spikeArray,
                                 label="Array spike source")
 
-ff_s = np.zeros(N_layer)
-lat_s = np.zeros(N_layer)
+ff_s = np.zeros(N_layer, dtype=np.uint)
+lat_s = np.zeros(N_layer, dtype=np.uint)
 
 existing_pre_ff = []
 existing_pre_lat = []
